+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-##### 0.10.0 - 25 June 2014
+##### 0.10.0 - 29 June 2014
 
 ###### Breaking API changes
 - #76 - Queries and filtering. See [TRANSITION.md](https://github.com/jmdobry/angular-data/blob/master/TRANSITION.md).
@@ -7,9 +7,11 @@
 
 ###### Backwards compatible API changes
 - #17 - Where predicates should be able to handle OR, not just AND
+- #23 - Computed Properties
 - #78 - Added optional callback to `bindOne` and `bindAll`
 - #79 - `ejectAll` should clear matching completed queries
 - #83 - Implement `DS.loadRelations(resourceName, instance(Id), relations[, options])`
+- #84 - idAttribute of a resource can be a computed property
 
 ##### 0.9.1 - 30 May 2014